chore: overriding table plugin []
